Group and Lie algebra elements

DecomposingRepresentations.AbstractGroupElem — Type

AbstractGroupElem

An abstract type representing a group element.

DecomposingRepresentations.AbstractLieAlgebraElem — Type

AbstractLieAlgebraElem

An abstract type representing an element of a Lie algebra.

GroupElem

DecomposingRepresentations.group — Method

group(::AbstractGroupElem) -> AbstractGroup

Returns the group to which the given group element belongs.

LieAlgebraElem

DecomposingRepresentations.algebra — Method

algebra(::AbstractLieAlgebraElem) -> AbstractLieAlgebra

Returns the Lie algebra to which the given Lie algebra element belongs.