Groups

DecomposingRepresentations.AbstractGroup — Type

AbstractGroup{T<:GroupType, F}

An abstract type representing a mathematical group. The type T represents a GroupType, while F represents the number field (or number type) over which the group is defined.

Group types

DecomposingRepresentations.GroupType — Type

GroupType

Abstract type representing a group type in the context of decomposing representations. This serves as a base type for all specific group types that will be defined.

DecomposingRepresentations.Finite — Type

Finite <: GroupType

Represents a finite group type. This type is used to categorize groups that have a finite number of elements.

DecomposingRepresentations.Lie — Type

Lie <: GroupType

Represents a Lie group type.

DecomposingRepresentations.Mixed — Type

Mixed <: GroupType

Represents a mixed group type. This type is used in direct products of finite groups with Lie groups.

Finite groups

Lie groups

DecomposingRepresentations.algebra — Method

algebra(::AbstractGroup{Lie}) -> AbstractLieAlgebra

Returns the Lie algebra of a given Lie group.

Lie algebras

DecomposingRepresentations.AbstractLieAlgebra — Type

AbstractLieAlgebra{F}

An abstract type representing a Lie algebra. The type F represents the number field (or number type) over which the Lie algebra is defined.

DecomposingRepresentations.name — Method

name(::AbstractLieAlgebra) -> String

Returns the name of the given Lie algebra.

DecomposingRepresentations.basis — Method

basis(::AbstractLieAlgebra)

Returns the basis of the given Lie algebra.

DecomposingRepresentations.chevalley_basis — Method

chevalley_basis(::AbstractLieAlgebra)

Returns the Chevalley basis of the given Lie algebra.

DecomposingRepresentations.dim — Method

dim(::AbstractLieAlgebra) -> Int

Returns the dimension of the given Lie algebra.

DecomposingRepresentations.rank — Method

rank(::AbstractLieAlgebra) -> Int

Returns the rank (i.e. the dimension of the Cartan subalgebra) of the given Lie algebra.

Direct product of groups

DecomposingRepresentations.AbstractDirectProductGroup — Type

AbstractDirectProductGroup{T<:GroupType, F} <: AbstractGroup{T, F}

An abstract type representing a direct product group.