Lie groups

DecomposingRepresentations.algebra — Method

algebra(::AbstractGroup{Lie}) -> AbstractLieAlgebra

Returns the Lie algebra of a given Lie group.

Lie algebras

DecomposingRepresentations.AbstractLieAlgebra — Type

AbstractLieAlgebra{F}

An abstract type representing a Lie algebra. The type F represents the number field (or number type) over which the Lie algebra is defined.

DecomposingRepresentations.name — Method

name(::AbstractLieAlgebra) -> String

Returns the name of the given Lie algebra.

DecomposingRepresentations.basis — Method

basis(::AbstractLieAlgebra)

Returns the basis of the given Lie algebra.

DecomposingRepresentations.chevalley_basis — Method

chevalley_basis(::AbstractLieAlgebra)

Returns the Chevalley basis of the given Lie algebra.

DecomposingRepresentations.dim — Method

dim(::AbstractLieAlgebra) -> Int

Returns the dimension of the given Lie algebra.

DecomposingRepresentations.rank — Method

rank(::AbstractLieAlgebra) -> Int

Returns the rank (i.e. the dimension of the Cartan subalgebra) of the given Lie algebra.