Types of groups

DecomposingGroupRepresentations.GroupType — Type

abstract type GroupType end

Abstract type representing a group type in the context of decomposing representations. The concrete types are Finite, Lie and Mixed.

DecomposingGroupRepresentations.Finite — Type

struct Finite <: GroupType end

Represents a finite group type. This type is used to categorize groups that have a finite number of elements.

DecomposingGroupRepresentations.Lie — Type

struct Lie <: GroupType end

Represents a Lie group type.

DecomposingGroupRepresentations.Mixed — Type

struct Mixed <: GroupType end

Represents a mixed group type. This type is used in direct products of finite groups with Lie groups.

DecomposingGroupRepresentations.AbstractGroup — Type

AbstractGroup{T<:GroupType, F}

An abstract type representing a reductive group. The type T represents a GroupType, while F represents the number field (or number type) over which the group is defined.