Group representations

DecomposingRepresentations.AbstractGroupRepresentation — Type

AbstractGroupRepresentation{T<:GroupType, S<:AbstractVectorSpace}

An abstract type representing a group representation. The type T represents a GroupType, and S represents an AbstractVectorSpace.

GroupRepresentation

IrreducibleRepresentation

IsotypicComponent

Methods

DecomposingRepresentations.action — Method

action(::AbstractGroupRepresentation) -> AbstractGroupAction

Returns the group action associated with the given group representation.

DecomposingRepresentations.group — Method

group(ρ::AbstractGroupRepresentation) -> AbstractGroup

Returns the group associated with the given group representation.

DecomposingRepresentations.space — Method

space(::AbstractGroupRepresentation) -> AbstractVectorSpace

Returns the vector space on which the given group representation acts.

DecomposingRepresentations.dim — Method

dim(ρ::AbstractGroupRepresentation) -> Int

Returns the dimension of the vector space on which the given group representation acts.

DecomposingRepresentations.irreducibles — Method

irreducibles(::AbstractGroupRepresentation) -> Any

Returns the irreducible components of the given group representation.

DecomposingRepresentations.isotypic_components — Method

isotypic_components(::AbstractGroupRepresentation) -> Any

Returns the isotypic components of the given group representation.