Vector spaces

DecomposingRepresentations.AbstractVectorSpace — Type

AbstractVectorSpace{F}

An abstract type representing a vector space. The type F represents the number field (or number type) over which the vector space is defined.

DecomposingRepresentations.basis — Method

basis(::AbstractVectorSpace)

Returns a basis of the given vector space.

DecomposingRepresentations.dim — Method

dim(::AbstractVectorSpace) -> Int

Returns the dimension of the given vector space.

VectorSpace

Direct sum

DecomposingRepresentations.AbstractDirectSum — Type

AbstractDirectSum{F} <: AbstractVectorSpace{F}

An abstract type representing a direct sum of vector spaces over the field (or number type) F.

DecomposingRepresentations.summands — Method

summands(::AbstractDirectSum)

Returns the summands of the given direct sum of vector spaces.

DecomposingRepresentations.nsummands — Method

nsummands(::AbstractDirectSum) -> Int

Returns the number of summands in the given direct sum of vector spaces.

Symmetric power

DecomposingRepresentations.AbstractSymmetricPower — Type

AbstractSymmetricPower{F} <: AbstractVectorSpace{F}

An abstract type representing a symmetric power of a vector space over the field (or number type) F.

DecomposingRepresentations.base_space — Function

base_space(::AbstractSymmetricPower) -> AbstractVectorSpace

Returns the base vector space of the given symmetric power. I.e. for $\mathrm{Sym}^n(V)$, this function returns $V$.

DecomposingRepresentations.power — Function

power(::AbstractSymmetricPower) -> Int

Returns the power of the given symmetric power.

Tensor product

DecomposingRepresentations.AbstractTensorProduct — Type

AbstractTensorProduct{F} <: AbstractVectorSpace{F}

An abstract type representing a tensor product of vector spaces over the field F.

DecomposingRepresentations.spaces — Function

spaces(::AbstractTensorProduct) -> Tuple{AbstractVectorSpace}

Returns the tuple of vector spaces that form the given tensor product.

HighestWeightModule